refactored node server structure
