+++ examples/server/node/api/me/index.js
@@ -1,0 +1,26 @@
+var express = require('express');
+var router = express.Router();
+
+var auth = require('../../auth/auth.service');
+var User = require('./user.model');
+
+router.get('/', auth.ensureAuthenticated, function(req, res) {
+  User.findById(req.user, function(err, user) {
+    res.send(user);
+  });
+});
+
+router.put('/', auth.ensureAuthenticated, function(req, res) {
+  User.findById(req.user, function(err, user) {
+    if (!user) {
+      return res.status(400).send({ message: 'User not found' });
+    }
+    user.displayName = req.body.displayName || user.displayName;
+    user.email = req.body.email || user.email;
+    user.save(function(err) {
+      res.status(200).end();
+    });
+  });
+});
+
+module.exports = router;
