+++ examples/server/node/auth/providers/github.js
@@ -1,0 +1,75 @@
+var express = require('express');
+var router = express.Router();
+var jwt = require('jwt-simple');
+var request = require('request');
+var qs = require('querystring');
+var config = require('../../config');
+var auth = require('../auth.service');
+var User = require('../../api/me/user.model');
+
+/*
+ |--------------------------------------------------------------------------
+ | Login with GitHub
+ |--------------------------------------------------------------------------
+ */
+router.post('/', function(req, res) {
+  var accessTokenUrl = 'https://github.com/login/oauth/access_token';
+  var userApiUrl = 'https://api.github.com/user';
+  var params = {
+    code: req.body.code,
+    client_id: req.body.clientId,
+    client_secret: config.GITHUB_SECRET,
+    redirect_uri: req.body.redirectUri
+  };
+
+  // Step 1. Exchange authorization code for access token.
+  request.get({ url: accessTokenUrl, qs: params }, function(err, response, accessToken) {
+    accessToken = qs.parse(accessToken);
+    var headers = { 'User-Agent': 'Satellizer' };
+
+    // Step 2. Retrieve profile information about the current user.
+    request.get({ url: userApiUrl, qs: accessToken, headers: headers, json: true }, function(err, response, profile) {
+
+      // Step 3a. Link user accounts.
+      if (req.headers.authorization) {
+        User.findOne({ github: profile.id }, function(err, existingUser) {
+          if (existingUser) {
+            return res.status(409).send({ message: 'There is already a GitHub account that belongs to you' });
+          }
+          var token = req.headers.authorization.split(' ')[1];
+          var payload = jwt.decode(token, config.TOKEN_SECRET);
+          User.findById(payload.sub, function(err, user) {
+            if (!user) {
+              return res.status(400).send({ message: 'User not found' });
+            }
+            user.github = profile.id;
+            user.picture = user.picture || profile.avatar_url;
+            user.displayName = user.displayName || profile.name;
+            user.save(function() {
+              var token = auth.createJWT(user);
+              res.send({ token: token });
+            });
+          });
+        });
+      } else {
+        // Step 3b. Create a new user account or return an existing one.
+        User.findOne({ github: profile.id }, function(err, existingUser) {
+          if (existingUser) {
+            var token = auth.createJWT(existingUser);
+            return res.send({ token: token });
+          }
+          var user = new User();
+          user.github = profile.id;
+          user.picture = profile.avatar_url;
+          user.displayName = profile.name;
+          user.save(function() {
+            var token = auth.createJWT(user);
+            res.send({ token: token });
+          });
+        });
+      }
+    });
+  });
+});
+
+module.exports = router;
