+++ examples/server/node/auth/providers/live.js
@@ -1,0 +1,78 @@
+var express = require('express');
+var router = express.Router();
+var jwt = require('jwt-simple');
+var request = require('request');
+var async = require('async');
+var config = require('../../config');
+var auth = require('../auth.service');
+var User = require('../../api/me/user.model');
+
+/*
+ |--------------------------------------------------------------------------
+ | Login with Windows Live
+ |--------------------------------------------------------------------------
+ */
+router.post('/', function(req, res) {
+  async.waterfall([
+    // Step 1. Exchange authorization code for access token.
+    function(done) {
+      var accessTokenUrl = 'https://login.live.com/oauth20_token.srf';
+      var params = {
+        code: req.body.code,
+        client_id: req.body.clientId,
+        client_secret: config.WINDOWS_LIVE_SECRET,
+        redirect_uri: req.body.redirectUri,
+        grant_type: 'authorization_code'
+      };
+      request.post(accessTokenUrl, { form: params, json: true }, function(err, response, accessToken) {
+        done(null, accessToken);
+      });
+    },
+    // Step 2. Retrieve profile information about the current user.
+    function(accessToken, done) {
+      var profileUrl = 'https://apis.live.net/v5.0/me?access_token=' + accessToken.access_token;
+      request.get({ url: profileUrl, json: true }, function(err, response, profile) {
+        done(err, profile);
+      });
+    },
+    function(profile) {
+      // Step 3a. Link user accounts.
+      if (!req.headers.authorization) {
+        User.findOne({ live: profile.id }, function(err, user) {
+          if (user) {
+            return res.send({ token: auth.createJWT(user) });
+          }
+          var newUser = new User();
+          newUser.live = profile.id;
+          newUser.displayName = profile.name;
+          newUser.save(function() {
+            var token = auth.createJWT(newUser);
+            res.send({ token: token });
+          });
+        });
+      } else {
+        // Step 3b. Create a new user or return an existing account.
+        User.findOne({ live: profile.id }, function(err, user) {
+          if (user) {
+            return res.status(409).send({ message: 'There is already a Windows Live account that belongs to you' });
+          }
+          var token = req.headers.authorization.split(' ')[1];
+          var payload = jwt.decode(token, config.TOKEN_SECRET);
+          User.findById(payload.sub, function(err, existingUser) {
+            if (!existingUser) {
+              return res.status(400).send({ message: 'User not found' });
+            }
+            existingUser.live = profile.id;
+            existingUser.displayName = existingUser.name;
+            existingUser.save(function() {
+              var token = auth.createJWT(existingUser);
+              res.send({ token: token });
+            });
+          });
+        });
+      }
+    }
+  ]);
+});
+
+module.exports = router;
