+++ examples/server/node/auth/providers/yahoo.js
@@ -1,0 +1,72 @@
+var express = require('express');
+var router = express.Router();
+var jwt = require('jwt-simple');
+var request = require('request');
+var config = require('../../config');
+var auth = require('../auth.service');
+var User = require('../../api/me/user.model');
+
+/*
+ |--------------------------------------------------------------------------
+ | Login with Yahoo
+ |--------------------------------------------------------------------------
+ */
+router.post('/auth/yahoo', function(req, res) {
+  var accessTokenUrl = 'https://api.login.yahoo.com/oauth2/get_token';
+  var clientId = req.body.clientId;
+  var clientSecret = config.YAHOO_SECRET;
+  var formData = {
+    code: req.body.code,
+    redirect_uri: req.body.redirectUri,
+    grant_type: 'authorization_code'
+  };
+  var headers = { Authorization: 'Basic ' + new Buffer(clientId + ':' + clientSecret).toString('base64') };
+
+  // Step 1. Exchange authorization code for access token.
+  request.post({ url: accessTokenUrl, form: formData, headers: headers, json: true }, function(err, response, body) {
+    var socialApiUrl = 'https://social.yahooapis.com/v1/user/' + body.xoauth_yahoo_guid + '/profile?format=json';
+    var headers = { Authorization: 'Bearer ' + body.access_token };
+
+    // Step 2. Retrieve profile information about the current user.
+    request.get({ url: socialApiUrl, headers: headers, json: true }, function(err, response, body) {
+
+      // Step 3a. Link user accounts.
+      if (req.headers.authorization) {
+        User.findOne({ yahoo: body.profile.guid }, function(err, existingUser) {
+          if (existingUser) {
+            return res.status(409).send({ message: 'There is already a Yahoo account that belongs to you' });
+          }
+          var token = req.headers.authorization.split(' ')[1];
+          var payload = jwt.decode(token, config.TOKEN_SECRET);
+          User.findById(payload.sub, function(err, user) {
+            if (!user) {
+              return res.status(400).send({ message: 'User not found' });
+            }
+            user.yahoo = body.profile.guid;
+            user.displayName = user.displayName || body.profile.nickname;
+            user.save(function() {
+              var token = auth.createJWT(user);
+              res.send({ token: token });
+            });
+          });
+        });
+      } else {
+        // Step 3b. Create a new user account or return an existing one.
+        User.findOne({ yahoo: body.profile.guid }, function(err, existingUser) {
+          if (existingUser) {
+            return res.send({ token: auth.createJWT(existingUser) });
+          }
+          var user = new User();
+          user.yahoo = body.profile.guid;
+          user.displayName = body.profile.nickname;
+          user.save(function() {
+            var token = auth.createJWT(user);
+            res.send({ token: token });
+          });
+        });
+      }
+    });
+  });
+});
+
+module.exports = router;
