+++ examples/server/node/package.json
@@ -6,6 +6,7 @@
     "async": "^0.9.2",
     "bcryptjs": "^2.1.0",
     "body-parser": "^1.12.4",
+    "colors": "^1.1.0",
     "express": "^4.12.4",
     "jwt-simple": "^0.3.0",
     "moment": "^2.10.3",
