+++ examples/server/node/routes.js
@@ -1,0 +1,4 @@
+module.exports = function(app) {
+  app.use('/auth', require('./auth'));
+  app.use('/api/me', require('./api/me'));
+};
